requirements.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.10.0
django-jalali-date==1.0.1
django-render-partial==0.4
djangorestframework==3.14.0
jdatetime==4.1.1
Pillow==9.4.0
pytz==2022.6
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tzdata==2023.3
